chore: Remove commented out code and update shopheader styling
The commented out code for importing the Shopheader component has been removed from the App.jsx file. Additionally, the styling of the Shopheader component has been updated to use a gray background color and add margin to the pageName text. This improves the visual appearance of the header in the shop section of the application.